ExpiringDate/requirements.txt
flask-restful
flask_restx
flask
pytesseract
numpy
opencv-python-headless
matplotlib
scikit-image
Pillow
dateparser
gunicorn
flask-cors
